+++ compiler/src/org.graalvm.compiler.lir.amd64/src/org/graalvm/compiler/lir/amd64/AMD64ArrayEqualsOp.java
@@ -138,16 +138,7 @@ public final class AMD64ArrayEqualsOp extends AMD64LIRInstruction {
 
         masm.movl(result, length); // copy
 
-        if (supportsAVX2(crb.target)) {
-            emitAVXCompare(crb, masm, result, array1, array2, length, trueLabel, falseLabel);
-        } else if (supportsSSE41(crb.target)) {
-            // this code is used for AVX as well because our backend correctly ensures that
-            // VEX-prefixed instructions are emitted if AVX is supported
-            emitSSE41Compare(crb, masm, result, array1, array2, length, trueLabel, falseLabel);
-        }
-
-        emit8ByteCompare(crb, masm, result, array1, array2, length, trueLabel, falseLabel);
-        emitTailCompares(masm, result, array1, array2, length, trueLabel, falseLabel);
+        emitArrayCompare(crb, masm, kind, result, array1, array2, length, temp4, temp5, tempXMM, vectorTemp1, vectorTemp2, trueLabel, falseLabel);
 
         // Return true
         masm.bind(trueLabel);
@@ -160,6 +151,21 @@ public final class AMD64ArrayEqualsOp extends AMD64LIRInstruction {
 
         // That's it
         masm.bind(done);
+    }
+
+    static void emitArrayCompare(CompilationResultBuilder crb, AMD64MacroAssembler masm, JavaKind kind,
+                    Register result, Register array1, Register array2, Register length,
+                    Value temp4, Value temp5, Value tempXMM, Value vectorTemp1, Value vectorTemp2,
+                    Label trueLabel, Label falseLabel) {
+        if (supportsAVX2(crb.target)) {
+            emitAVXCompare(crb, masm, kind, result, array1, array2, length, temp4, temp5, tempXMM, vectorTemp1, vectorTemp2, trueLabel, falseLabel);
+        } else if (supportsSSE41(crb.target)) {
+            // this code is used for AVX as well because our backend correctly ensures that
+            // VEX-prefixed instructions are emitted if AVX is supported
+            emitSSE41Compare(crb, masm, kind, result, array1, array2, length, temp4, temp5, tempXMM, vectorTemp1, vectorTemp2, trueLabel, falseLabel);
+        }
+        emit8ByteCompare(crb, masm, kind, result, array1, array2, length, temp4, tempXMM, trueLabel, falseLabel);
+        emitTailCompares(masm, kind, result, array1, array2, length, temp4, tempXMM, trueLabel, falseLabel);
     }
 
     /**
@@ -181,11 +187,14 @@ public final class AMD64ArrayEqualsOp extends AMD64LIRInstruction {
     /**
      * Emits code that uses SSE4.1 128-bit (16-byte) vector compares.
      */
-    private void emitSSE41Compare(CompilationResultBuilder crb, AMD64MacroAssembler masm, Register result, Register array1, Register array2, Register length, Label trueLabel, Label falseLabel) {
+    private static void emitSSE41Compare(CompilationResultBuilder crb, AMD64MacroAssembler masm, JavaKind kind,
+                    Register result, Register array1, Register array2, Register length,
+                    Value temp4, Value temp5, Value tempXMM, Value vectorTemp1, Value vectorTemp2,
+                    Label trueLabel, Label falseLabel) {
         assert supportsSSE41(crb.target);
 
-        Register vector1 = asRegister(vectorTemp1, AMD64Kind.DOUBLE);
-        Register vector2 = asRegister(vectorTemp2, AMD64Kind.DOUBLE);
+        Register vector1 = asRegister(vectorTemp1);
+        Register vector2 = asRegister(vectorTemp2);
 
         Label loop = new Label();
         Label compareTail = new Label();
@@ -223,7 +232,7 @@ public final class AMD64ArrayEqualsOp extends AMD64LIRInstruction {
             Label unalignedCheck = new Label();
             masm.jmpb(unalignedCheck);
             masm.bind(nanCheck);
-            emitFloatCompareWithinRange(crb, masm, array1, array2, length, 0, falseLabel, SSE4_1_VECTOR_SIZE);
+            emitFloatCompareWithinRange(crb, masm, kind, array1, array2, length, temp4, temp5, tempXMM, 0, falseLabel, SSE4_1_VECTOR_SIZE);
             masm.jmpb(loopCheck);
             masm.bind(unalignedCheck);
         }
@@ -238,7 +247,7 @@ public final class AMD64ArrayEqualsOp extends AMD64LIRInstruction {
         masm.ptest(vector1, vector1);
         if (requiresNaNCheck) {
             masm.jcc(ConditionFlag.Zero, trueLabel);
-            emitFloatCompareWithinRange(crb, masm, array1, array2, result, -SSE4_1_VECTOR_SIZE, falseLabel, SSE4_1_VECTOR_SIZE);
+            emitFloatCompareWithinRange(crb, masm, kind, array1, array2, result, temp4, temp5, tempXMM, -SSE4_1_VECTOR_SIZE, falseLabel, SSE4_1_VECTOR_SIZE);
         } else {
             masm.jcc(ConditionFlag.NotZero, falseLabel);
         }
@@ -264,11 +273,14 @@ public final class AMD64ArrayEqualsOp extends AMD64LIRInstruction {
      */
     private static final int AVX_VECTOR_SIZE = 32;
 
-    private void emitAVXCompare(CompilationResultBuilder crb, AMD64MacroAssembler masm, Register result, Register array1, Register array2, Register length, Label trueLabel, Label falseLabel) {
+    private static void emitAVXCompare(CompilationResultBuilder crb, AMD64MacroAssembler masm, JavaKind kind, Register result,
+                    Register array1, Register array2, Register length,
+                    Value temp4, Value temp5, Value tempXMM, Value vectorTemp1, Value vectorTemp2,
+                    Label trueLabel, Label falseLabel) {
         assert supportsAVX2(crb.target);
 
-        Register vector1 = asRegister(vectorTemp1, AMD64Kind.DOUBLE);
-        Register vector2 = asRegister(vectorTemp2, AMD64Kind.DOUBLE);
+        Register vector1 = asRegister(vectorTemp1);
+        Register vector2 = asRegister(vectorTemp2);
 
         Label loop = new Label();
         Label compareTail = new Label();
@@ -306,7 +318,7 @@ public final class AMD64ArrayEqualsOp extends AMD64LIRInstruction {
             Label unalignedCheck = new Label();
             masm.jmpb(unalignedCheck);
             masm.bind(nanCheck);
-            emitFloatCompareWithinRange(crb, masm, array1, array2, length, 0, falseLabel, AVX_VECTOR_SIZE);
+            emitFloatCompareWithinRange(crb, masm, kind, array1, array2, length, temp4, temp5, tempXMM, 0, falseLabel, AVX_VECTOR_SIZE);
             masm.jmpb(loopCheck);
             masm.bind(unalignedCheck);
         }
@@ -321,7 +333,7 @@ public final class AMD64ArrayEqualsOp extends AMD64LIRInstruction {
         masm.vptest(vector1, vector1);
         if (requiresNaNCheck) {
             masm.jcc(ConditionFlag.Zero, trueLabel);
-            emitFloatCompareWithinRange(crb, masm, array1, array2, result, -AVX_VECTOR_SIZE, falseLabel, AVX_VECTOR_SIZE);
+            emitFloatCompareWithinRange(crb, masm, kind, array1, array2, result, temp4, temp5, tempXMM, -AVX_VECTOR_SIZE, falseLabel, AVX_VECTOR_SIZE);
         } else {
             masm.jcc(ConditionFlag.NotZero, falseLabel);
         }
@@ -339,7 +351,8 @@ public final class AMD64ArrayEqualsOp extends AMD64LIRInstruction {
     /**
      * Emits code that uses 8-byte vector compares.
      */
-    private void emit8ByteCompare(CompilationResultBuilder crb, AMD64MacroAssembler masm, Register result, Register array1, Register array2, Register length, Label trueLabel, Label falseLabel) {
+    private static void emit8ByteCompare(CompilationResultBuilder crb, AMD64MacroAssembler masm, JavaKind kind, Register result, Register array1, Register array2, Register length, Value temp4,
+                    Value tempXMM, Label trueLabel, Label falseLabel) {
         Label loop = new Label();
         Label compareTail = new Label();
 
@@ -378,7 +391,7 @@ public final class AMD64ArrayEqualsOp extends AMD64LIRInstruction {
             masm.bind(nanCheck);
             // At most two iterations, unroll in the emitted code.
             for (int offset = 0; offset < VECTOR_SIZE; offset += kind.getByteCount()) {
-                emitFloatCompare(masm, array1, array2, length, offset, falseLabel, kind.getByteCount() == VECTOR_SIZE);
+                emitFloatCompare(masm, kind, array1, array2, length, temp4, tempXMM, offset, falseLabel, kind.getByteCount() == VECTOR_SIZE);
             }
             masm.jmpb(loopCheck);
             masm.bind(unalignedCheck);
@@ -394,7 +407,7 @@ public final class AMD64ArrayEqualsOp extends AMD64LIRInstruction {
             masm.jcc(ConditionFlag.Equal, trueLabel);
             // At most two iterations, unroll in the emitted code.
             for (int offset = 0; offset < VECTOR_SIZE; offset += kind.getByteCount()) {
-                emitFloatCompare(masm, array1, array2, result, -VECTOR_SIZE + offset, falseLabel, kind.getByteCount() == VECTOR_SIZE);
+                emitFloatCompare(masm, kind, array1, array2, result, temp4, tempXMM, -VECTOR_SIZE + offset, falseLabel, kind.getByteCount() == VECTOR_SIZE);
             }
         } else {
             masm.jccb(ConditionFlag.NotEqual, falseLabel);
@@ -408,7 +421,8 @@ public final class AMD64ArrayEqualsOp extends AMD64LIRInstruction {
     /**
      * Emits code to compare the remaining 1 to 4 bytes.
      */
-    private void emitTailCompares(AMD64MacroAssembler masm, Register result, Register array1, Register array2, Register length, Label trueLabel, Label falseLabel) {
+    private static void emitTailCompares(AMD64MacroAssembler masm, JavaKind kind, Register result, Register array1, Register array2, Register length, Value temp4, Value tempXMM,
+                    Label trueLabel, Label falseLabel) {
         Label compare2Bytes = new Label();
         Label compare1Byte = new Label();
 
@@ -422,7 +436,7 @@ public final class AMD64ArrayEqualsOp extends AMD64LIRInstruction {
             masm.cmpl(temp, new AMD64Address(array2, 0));
             if (kind == JavaKind.Float) {
                 masm.jccb(ConditionFlag.Equal, trueLabel);
-                emitFloatCompare(masm, array1, array2, Register.None, 0, falseLabel, true);
+                emitFloatCompare(masm, kind, array1, array2, Register.None, temp4, tempXMM, 0, falseLabel, true);
                 masm.jmpb(trueLabel);
             } else {
                 masm.jccb(ConditionFlag.NotEqual, falseLabel);
@@ -467,7 +481,7 @@ public final class AMD64ArrayEqualsOp extends AMD64LIRInstruction {
     /**
      * Emits code to fall through if {@code src} is NaN, otherwise jump to {@code branchOrdered}.
      */
-    private void emitNaNCheck(AMD64MacroAssembler masm, AMD64Address src, Label branchIfNonNaN) {
+    private static void emitNaNCheck(AMD64MacroAssembler masm, JavaKind kind, Value tempXMM, AMD64Address src, Label branchIfNonNaN) {
         assert kind.isNumericFloat();
         Register tempXMMReg = asRegister(tempXMM);
         if (kind == JavaKind.Float) {
@@ -482,7 +496,8 @@ public final class AMD64ArrayEqualsOp extends AMD64LIRInstruction {
     /**
      * Emits code to compare if two floats are bitwise equal or both NaN.
      */
-    private void emitFloatCompare(AMD64MacroAssembler masm, Register base1, Register base2, Register index, int offset, Label falseLabel, boolean skipBitwiseCompare) {
+    private static void emitFloatCompare(AMD64MacroAssembler masm, JavaKind kind, Register base1, Register base2, Register index, Value temp4, Value tempXMM, int offset, Label falseLabel,
+                    boolean skipBitwiseCompare) {
         AMD64Address address1 = new AMD64Address(base1, index, Scale.Times1, offset);
         AMD64Address address2 = new AMD64Address(base2, index, Scale.Times1, offset);
 
@@ -502,8 +517,8 @@ public final class AMD64ArrayEqualsOp extends AMD64LIRInstruction {
             masm.jccb(ConditionFlag.Equal, bitwiseEqual);
         }
 
-        emitNaNCheck(masm, address1, falseLabel);
-        emitNaNCheck(masm, address2, falseLabel);
+        emitNaNCheck(masm, kind, tempXMM, address1, falseLabel);
+        emitNaNCheck(masm, kind, tempXMM, address2, falseLabel);
 
         masm.bind(bitwiseEqual);
     }
@@ -511,7 +526,8 @@ public final class AMD64ArrayEqualsOp extends AMD64LIRInstruction {
     /**
      * Emits code to compare float equality within a range.
      */
-    private void emitFloatCompareWithinRange(CompilationResultBuilder crb, AMD64MacroAssembler masm, Register base1, Register base2, Register index, int offset, Label falseLabel, int range) {
+    private static void emitFloatCompareWithinRange(CompilationResultBuilder crb, AMD64MacroAssembler masm, JavaKind kind, Register base1, Register base2, Register index, Value temp4, Value temp5,
+                    Value tempXMM, int offset, Label falseLabel, int range) {
         assert kind.isNumericFloat();
         Label loop = new Label();
         Register i = asRegister(temp5);
@@ -521,7 +537,7 @@ public final class AMD64ArrayEqualsOp extends AMD64LIRInstruction {
         // Align the main loop
         masm.align(crb.target.wordSize * 2);
         masm.bind(loop);
-        emitFloatCompare(masm, base1, base2, index, offset, falseLabel, kind.getByteCount() == range);
+        emitFloatCompare(masm, kind, base1, base2, index, temp4, tempXMM, offset, falseLabel, kind.getByteCount() == range);
         masm.addq(index, kind.getByteCount());
         masm.addq(i, kind.getByteCount());
         masm.jccb(ConditionFlag.NotZero, loop);