+++ compiler/src/org.graalvm.compiler.replacements.amd64/src/org/graalvm/compiler/replacements/amd64/AMD64StringIndexOfStringNode.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2018, 2018, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -24,11 +24,10 @@
  */
 package org.graalvm.compiler.replacements.amd64;
 
-import static org.graalvm.compiler.nodeinfo.NodeSize.SIZE_64;
-
+import jdk.vm.ci.meta.JavaKind;
+import jdk.vm.ci.meta.Value;
 import org.graalvm.compiler.core.common.type.StampFactory;
 import org.graalvm.compiler.graph.NodeClass;
-import org.graalvm.compiler.graph.NodeInputList;
 import org.graalvm.compiler.nodeinfo.InputType;
 import org.graalvm.compiler.nodeinfo.NodeCycles;
 import org.graalvm.compiler.nodeinfo.NodeInfo;
@@ -43,51 +42,48 @@ import org.graalvm.compiler.nodes.spi.NodeLIRBuilderTool;
 import org.graalvm.word.LocationIdentity;
 import org.graalvm.word.Pointer;
 
-import jdk.vm.ci.meta.JavaKind;
-import jdk.vm.ci.meta.Value;
+import static org.graalvm.compiler.nodeinfo.NodeSize.SIZE_512;
 
-@NodeInfo(size = SIZE_64, cycles = NodeCycles.CYCLES_UNKNOWN)
-public class AMD64StringIndexOfNode extends FixedWithNextNode implements LIRLowerable, MemoryAccess {
-    public static final NodeClass<AMD64StringIndexOfNode> TYPE = NodeClass.create(AMD64StringIndexOfNode.class);
+@NodeInfo(size = SIZE_512, cycles = NodeCycles.CYCLES_UNKNOWN)
+public class AMD64StringIndexOfStringNode extends FixedWithNextNode implements LIRLowerable, MemoryAccess {
 
-    @OptionalInput(InputType.Memory) protected MemoryNode lastLocationAccess;
+    public static final NodeClass<AMD64StringIndexOfStringNode> TYPE = NodeClass.create(AMD64StringIndexOfStringNode.class);
 
-    @Input protected NodeInputList<ValueNode> arguments;
+    private final JavaKind kind;
 
-    public AMD64StringIndexOfNode(ValueNode sourcePointer, ValueNode sourceCount, ValueNode targetPointer, ValueNode targetCount) {
-        super(TYPE, StampFactory.forInteger(32));
-        this.arguments = new NodeInputList<>(this, new ValueNode[]{sourcePointer, sourceCount, targetPointer, targetCount});
+    @Input private ValueNode haystackPointer;
+    @Input private ValueNode haystackLength;
+    @Input private ValueNode needlePointer;
+    @Input private ValueNode needleLength;
+
+    @OptionalInput(InputType.Memory) private MemoryNode lastLocationAccess;
+
+    public AMD64StringIndexOfStringNode(@ConstantNodeParameter JavaKind kind,
+                    ValueNode haystackPointer,
+                    ValueNode haystackLength,
+                    ValueNode needlePointer,
+                    ValueNode needleLength) {
+        super(TYPE, StampFactory.forKind(JavaKind.Int));
+        this.kind = kind;
+        this.haystackPointer = haystackPointer;
+        this.haystackLength = haystackLength;
+        this.needlePointer = needlePointer;
+        this.needleLength = needleLength;
     }
 
     @Override
     public LocationIdentity getLocationIdentity() {
-        return NamedLocationIdentity.getArrayLocation(JavaKind.Char);
-    }
-
-    ValueNode sourcePointer() {
-        return arguments.get(0);
-    }
-
-    ValueNode sourceCount() {
-        return arguments.get(1);
-    }
-
-    ValueNode targetPointer() {
-        return arguments.get(2);
-    }
-
-    ValueNode targetCount() {
-        return arguments.get(3);
+        return NamedLocationIdentity.getArrayLocation(kind);
     }
 
     @Override
     public void generate(NodeLIRBuilderTool gen) {
-        int constantTargetCount = -1;
-        if (targetCount().isConstant()) {
-            constantTargetCount = targetCount().asJavaConstant().asInt();
+        int constantNeedleLength = -1;
+        if (needleLength.isConstant()) {
+            constantNeedleLength = needleLength.asJavaConstant().asInt();
         }
-        Value result = gen.getLIRGeneratorTool().emitStringIndexOf(gen.operand(sourcePointer()), gen.operand(sourceCount()), gen.operand(targetPointer()), gen.operand(targetCount()),
-                        constantTargetCount);
+        Value result = gen.getLIRGeneratorTool().emitStringIndexOfString(kind,
+                        gen.operand(haystackPointer), gen.operand(haystackLength), gen.operand(needlePointer), gen.operand(needleLength), constantNeedleLength);
         gen.setResult(this, result);
     }
 
@@ -103,5 +99,5 @@ public class AMD64StringIndexOfNode extends FixedWithNextNode implements LIRLowe
     }
 
     @NodeIntrinsic
-    public static native int optimizedStringIndexPointer(Pointer sourcePointer, int sourceCount, Pointer targetPointer, int targetCount);
+    public static native int optimizedStringIndexOf(@ConstantNodeParameter JavaKind kind, Pointer haystackPointer, int haystackLength, Pointer needlePointer, int needleLength);
 }