+++ truffle/CHANGELOG.md
@@ -50,6 +50,7 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 * Debugger API methods now throw [DebugException](http://www.graalvm.org/truffle/javadoc/com/oracle/truffle/api/debug/DebugException.html) on language failures.
 * Deprecated API methods that use `java.beans` package in [AllocationReporter](http://www.graalvm.org/truffle/javadoc/com/oracle/truffle/api/instrumentation/AllocationReporter.html) and [Debugger](http://www.graalvm.org/truffle/javadoc/com/oracle/truffle/api/debug/Debugger.html). New add/remove listener methods were introduced as a replacement.
 * [FrameDescriptor](http://www.graalvm.org/truffle/javadoc/com/oracle/truffle/api/frame/FrameDescriptor.html) no longer shares a lock with a RootNode.
+* Added `ArrayUtils` API providing additional array and/or string operations that may be intrinsified by the compiler.
 
 ## Version 1.0.0 RC5